+++ Tests/Unit/Utility/ContentObjectUtilityTest.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace T3v\T3vCore\Tests\Unit\Utility;
 
-use Nimut\TestingFramework\TestCase\UnitTestCase;
 use T3v\T3vCore\Utility\ContentObjectUtility;
+use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 
 /**
  * The content object utility test class.