+++ Tests/Unit/Utility/IconUtilityTest.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace T3v\T3vCore\Tests\Unit\Utility;
 
-use Nimut\TestingFramework\TestCase\UnitTestCase;
 use T3v\T3vCore\Utility\IconUtility;
+use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 
 /**
  * The icon utility test class.